+++ tftf/tests/runtime_services/secure_service/test_ffa_setup_and_discovery.c
@@ -98,8 +98,7 @@ test_result_t test_ffa_features(void)
 			FFA_VERSION_1_1},
 		{"FFA_YIELD_32", FFA_MSG_YIELD, FFA_ERROR},
 	};
-	unsigned int test_target_size =
-		get_ffa_feature_test_target(&func_ids_target);
+	size_t test_target_size = get_ffa_feature_test_target(&func_ids_target);
 
 	SKIP_TEST_IF_FFA_VERSION_LESS_THAN(1, 0);
